a lot of progress

+++ AMTProject01/src/main/java/Services/Team/TeamManagerSQL.java
@@ -38,13 +38,13 @@ public class TeamManagerSQL implements TeamManager{
     }
     
     @Override
-    public List<Team> getAllTeams(){
+    public List<Team> getAllTeams(int creator){
         
             ArrayList<Team> teams = new ArrayList();
 
         try {
         Connection connection = dataSource.getConnection();
-        PreparedStatement pstmt = connection.prepareStatement("SELECT * FROM `Team`");
+        PreparedStatement pstmt = connection.prepareStatement("SELECT * FROM `Team` WHERE creator_id="+creator);
         ResultSet rs = pstmt.executeQuery();
         while (rs.next()) {
           int id = rs.getInt("team_id");
@@ -61,14 +61,14 @@ public class TeamManagerSQL implements TeamManager{
   }
        
     @Override
-    public Team getTeam(String name){
+    public Team getTeam(String name,int creator){
         
         int id = 0;
         String team ="";
         
          try {
         Connection connection = dataSource.getConnection();
-        PreparedStatement pstmt = connection.prepareStatement("SELECT * FROM Team WHERE name = '"+name+"'");
+        PreparedStatement pstmt = connection.prepareStatement("SELECT * FROM Team WHERE creator_id="+creator+" AND name = '"+name+"'");
         ResultSet rs = pstmt.executeQuery();
         while (rs.next()) {
           id = rs.getInt("team_id");
@@ -85,13 +85,13 @@ public class TeamManagerSQL implements TeamManager{
     }
     
     @Override
-    public ArrayList<Team> getTeamPartial(String name){
+    public ArrayList<Team> getTeamPartial(String name,int creator){
         
         ArrayList<Team> teams = new ArrayList();
         
          try {
         Connection connection = dataSource.getConnection();
-        PreparedStatement pstmt = connection.prepareStatement("SELECT * FROM Team WHERE name LIKE '"+name+"%'");
+        PreparedStatement pstmt = connection.prepareStatement("SELECT * FROM Team WHERE creator_id="+creator+" AND name LIKE '"+name+"%'");
         ResultSet rs = pstmt.executeQuery();
         while (rs.next()) {
           int id = rs.getInt("team_id");
@@ -99,11 +99,12 @@ public class TeamManagerSQL implements TeamManager{
           teams.add(new Team(id,team));
         }
         pstmt.close();
-      
-    } catch (SQLException ex) {
-      Logger.getLogger(TeamManagerSQL.class.getName()).log(Level.SEVERE, null, ex);
-    }
-    return  teams;
+              connection.close();
+
+        } catch (SQLException ex) {
+          Logger.getLogger(TeamManagerSQL.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        return  teams;
 
     }
     
@@ -111,12 +112,12 @@ public class TeamManagerSQL implements TeamManager{
     
     
     @Override
-    public int getNumberOfTeeams(){
+    public int getNumberOfTeams(int creator ){
         
         int number = 0;
         try {
             Connection connection = dataSource.getConnection();
-            PreparedStatement pstmt = connection.prepareStatement("SELECT COUNT(*) AS count FROM Team");
+            PreparedStatement pstmt = connection.prepareStatement("SELECT COUNT(*) AS count FROM Team WHERE creator_id="+creator);
               ResultSet rs = pstmt.executeQuery();
         while (rs.next()) {
            number = rs.getInt("count");
@@ -150,13 +151,13 @@ public class TeamManagerSQL implements TeamManager{
     
     
     @Override
-    public void deleteTeam(Team t){
+    public void deleteTeam(Team t, int creator){
         
         try {
         Connection connection = dataSource.getConnection();
         
         
-        PreparedStatement pstmt = connection.prepareStatement("DELETE FROM Team WHERE team_id="+t.getId());
+        PreparedStatement pstmt = connection.prepareStatement("DELETE FROM Team WHERE creator_id="+creator+" AND team_id="+t.getId());
          pstmt.execute();
        
         pstmt.close();
